XlsxToPdf {XlsxToPdf}	R Documentation

XlsxToPdf

Description

Essa função foi criada no intuito de facilitar a transformação de arquivos no formato xlsx em pdf. É gerado um script vbs que abre o arquivo indicado e salva como pdf.

Usage

XlsxToPdf(x = "MeuArquivo.xlsx", Sheet = "Plan1", Output = "MeuArquivo.pdf")

Examples

XlsxToPdf(x = "MeuArquivo.xlsx", Sheet = "Plan1", Output = "MeuArquivo.pdf")

[Package XlsxToPdf version 0.1.0 Index]